adapt to the deletion of pkgCacheFile::Close() in ALT's APT
(In ALT, it was deleted, because it's easier to extend just the dtor
in subclasses.)

Since a pkgCacheFile object is intended to be one-shot (lazy
immutable, i.e., it internally constructs the owned objects and then
always gives the same value), there is not much sense in having a
separate Close() method. If we follow the immutability logic, after
Close() it is not valid anymore.

(Unfortunately, this is a change in the API of ALT's APT. Here is a
justification for this and a few subsequent changes.

The old behavior of pkgCacheFile's methods in ALT's APT didn't match
the expectation of modern clients, such as PackageKit.

And when correcting the semantics of the pkgCacheFile class to match
the modern one, I rewrote it in a way that really guarantees that no
one will be able to change the owned pointers, and also I dared to
make it more simple and clear: the pointers to owned objects returned
from an instance of pkgCacheFile will simply never change during its
life; before, with Close(), the returned pointers would be new after
Close(), which is not really needed in actual applications, as we see.

After the rewrite, the use of pkgCacheFile becomes alike to the common
RAII idiom[1], with the specific feature that the resources are acquired
lazily after construction.

Open() has been left as a way to force the acquisition of all
"resources", i.e., the build of all caches, and to check errors during it.

[1]: https://en.cppreference.com/w/cpp/language/raii "Classes with
open()/close(), lock()/unlock(), or init()/copyFrom()/destroy() member
functions are typical examples of non-RAII classes"

)